# openapi/openapi_carizon.yaml
openapi: 3.0.3
info:
  title: Carizon API
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /cars:
    get:
      summary: Search & list vehicles
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: priceMin
          schema:
            type: integer
        - in: query
          name: priceMax
          schema:
            type: integer
        - in: query
          name: kmMax
          schema:
            type: integer
        - in: query
          name: yymmFrom
          schema:
            type: integer
        - in: query
          name: yymmTo
          schema:
            type: integer
        - in: query
          name: fuel
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: transmission
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: body
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: region
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - price_asc
              - km_asc
              - latest
              - value_desc
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: OK
  /cars/{id}:
    get:
      summary: Vehicle detail + cross-platform comparison
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /cars/{id}/history:
    get:
      summary: Price & visibility history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: CARIZON_SESSION
security:
  - sessionCookie: []